Generalize the type of the size_type for containers. Use allocator traits to infer such types. Default usize is a u32

+++ performance/perf/memory.cpp
@@ -201,7 +201,7 @@ void RecordStackAllocator(const AllocationSettings &p_Settings) noexcept
     DynamicArray<ExampleData *> allocated{p_Settings.MaxPasses};
     file << "passes,stack_alloc (ns),stack_dealloc (ns)\n";
 
-    StackAllocator allocator{p_Settings.MaxPasses * sizeof(ExampleData)};
+    StackAllocator allocator{p_Settings.MaxPasses * static_cast<usize>(sizeof(ExampleData))};
     for (usize passes = p_Settings.MinPasses; passes <= p_Settings.MaxPasses; passes += p_Settings.PassIncrement)
     {
         Clock clock;
